Properly fix for TB2 using event handlers
Wrapping all the TB2 methods is slow and dangerous.  Instead...

At the start of event processing...
1. lock the history storage
2. read in a stored history object
3. Use that history object

Process the event normally

At the end of event processing...
1. Write out the history object to the storage file
2. Unlock the history storage

This synchronizes not just the test number, but all events
between processes, not matter if they're using Test::Builder
or not.

All but a few tests pass with TB2.