Fix basic subtest handling
When a subtest happens, create a new pair of handlers.  Had to do some special work
to make sure the new early and late handlers are paired even though they are
created separately.

This is not a complete subtest solution, multiple processes cannot be in their own
subtests at the same time.